OCR and Face Recog Project
Uses PIL to deal with images, uses OpenCV to use the Haar Classifier and Pytesseract for OCR. The project searches for the keyword specified by the user and searches the images for the keyword. If the keyword is found then it displays the faces in the corresponding image if there are any.